Add a common testing library
This implements a smaller helper library for performing tests of array
sorting functions, and refactors the written examples so as to use these
functions.

This involved unifying the interface of all of the sorting functions, so
that they accept only one argument (an array of integers) and return a
sorted version of this array. In the case of the insertion sort, this
involved refactoring the existing implementation to remove the
compare_func, and also make it in-place.